separate route for raw files
